requirements.txt
requests==2.33.1
beautifulsoup4==4.12.3
google-generativeai==0.8.6
groq==1.1.2
lxml==5.2.2
